fix: show open .env message always
